requirements.txt
astroid==2.4.2
autopep8==1.5.4
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
dnspython==2.1.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.10.0
mccabe==0.6.1
pep8==1.7.1
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyDocX==0.9.10
PyJWT==2.0.1
pylint==2.6.0
pymongo==3.11.2
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
Unidecode==1.1.2
Werkzeug==1.0.1
wrapt==1.12.1
